Create LICENSE
I'm interested in using some code from here, but it's currently not published under any license, meaning it can't be used in any other projects. This is a pull request to make it available under the MIT license. of course, there's no obligation to add one, but since this is GitHub and this is a 'lib', I presume the omission isn't intentional? thanks in advance. paging @nmbook, since they're the only contributor